requirements.txt
aiogram==2.23.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-dramatiq==0.11.5
django-resized==1.0.2
dramatiq==1.14.2
eskiz-sms==0.2.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fonttools==4.42.1
fpdf2==2.7.5
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
numpy==1.25.2
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pika==1.3.2
Pillow==10.0.0
prometheus-client==0.17.1
psycopg2-binary==2.9.7
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
